requirements.txt
python-dotenv
pinecone-client[grpc]
sentence-transformers
langchain-groq
langchain-community
langgraph
gradio